Revision 2: Optimize entityMap computation and make parseEntityReferences fully generic
